brilliant-api/requirements.txt
fastapi==0.111.0
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.7.4
pytest==8.3.3
httpx==0.27.0
pydantic-settings==2.6.1
tortoise-orm==0.21.7
aerich==0.7.2
asyncpg==0.30.0